# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MobileCRM API Documentation",
        "description": "MobileCRM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/business/hotel/rooms": {
            "get": {
                "tags": [
                    "Hotel - Rooms"
                ],
                "summary": "List Hotel Rooms",
                "description": "Returns all rooms for the active business, optionally filtered.",
                "operationId": "38313836b7984412add95ac803baeca5",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "room_type_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "floor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rooms retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hotel - Rooms"
                ],
                "summary": "Create Room",
                "operationId": "6e6201893ddf824a7d608a5fa3090378",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_number",
                                    "room_type_id"
                                ],
                                "properties": {
                                    "room_number": {
                                        "type": "string"
                                    },
                                    "room_type_id": {
                                        "type": "integer"
                                    },
                                    "floor": {
                                        "type": "string"
                                    },
                                    "is_ac": {
                                        "type": "boolean"
                                    },
                                    "current_tariff": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/rooms/{id}": {
            "get": {
                "tags": [
                    "Hotel - Rooms"
                ],
                "summary": "Get Room Details",
                "operationId": "3b98a9e0da40553c4d8b4881e5ad059a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotel - Rooms"
                ],
                "summary": "Update Room",
                "operationId": "90a2999a9abb9dc70370f30571b3be7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotel - Rooms"
                ],
                "summary": "Delete Room",
                "operationId": "8cd30b244e07c78af19eef128a184954",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/rooms/{id}/status": {
            "patch": {
                "tags": [
                    "Hotel - Rooms"
                ],
                "summary": "Update Room Status",
                "operationId": "11f6b4acea493389c50c5732c22ae327",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer Token to authenticate",
                "scheme": "bearer"
            }
        }
    },
...
}